+++ ultrasonic/src/main/kotlin/org/moire/ultrasonic/di/ApplicationModule.kt
@@ -4,6 +4,7 @@ import org.koin.dsl.module
 import org.moire.ultrasonic.data.ActiveServerProvider
 import org.moire.ultrasonic.subsonic.ImageLoaderProvider
 import org.moire.ultrasonic.util.CacheCleaner
+import org.moire.ultrasonic.util.MediaDeviceExporter
 
 /**
  * This Koin module contains the registration of general classes needed for Ultrasonic
@@ -12,4 +13,5 @@ val applicationModule = module {
     single { ActiveServerProvider(get()) }
     single { ImageLoaderProvider() }
     single { CacheCleaner() }
+    single { MediaDeviceExporter(get()) }
 }